# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Wash Booking App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export type RegisterUserBodyRole =
  (typeof RegisterUserBodyRole)[keyof typeof RegisterUserBodyRole];

export const RegisterUserBodyRole = {
  customer: "customer",
  washer: "washer",
  owner: "owner",
} as const;

export interface RegisterUserBody {
  name: string;
  email: string;
  password: string;
  role: RegisterUserBodyRole;
  address: string;
  city: string;
  /** @nullable */
  phone?: string | null;
}

export interface LoginUserBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  customer: "customer",
  washer: "washer",
  owner: "owner",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  address: string;
  city: string;
  /** @nullable */
  phone?: string | null;
  createdAt: string;
}

export interface CarwashOwner {
  id: number;
  userId: number;
  businessName: string;
  address: string;
  city: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  description?: string | null;
  ownerName: string;
  createdAt: string;
}

export interface CreateCarwashOwnerBody {
  businessName: string;
  address: string;
  city: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface Washer {
  id: number;
  userId: number;
  name: string;
  address: string;
  city: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  pricePerWash: number;
  available: boolean;
  /** @nullable */
  rating?: number | null;
  totalWashes: number;
  createdAt: string;
}

export interface CreateWasherBody {
  address: string;
  city: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  pricePerWash: number;
}

export interface UpdateWasherBody {
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  pricePerWash?: number | null;
  /** @nullable */
  available?: boolean | null;
}

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  accepted: "accepted",
  declined: "declined",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Booking {
  id: number;
  customerId: number;
  washerId: number;
  customerName: string;
  washerName: string;
  customerAddress: string;
  scheduledAt: string;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  priceQuoted: number;
  createdAt: string;
}

export interface CreateBookingBody {
  washerId: number;
  customerAddress: string;
  scheduledAt: string;
  /** @nullable */
  notes?: string | null;
}

export type ListWashersParams = {
  available?: boolean;
};

export type ListBookingsParams = {
  role?: ListBookingsRole;
};

export type ListBookingsRole =
  (typeof ListBookingsRole)[keyof typeof ListBookingsRole];

export const ListBookingsRole = {
  customer: "customer",
  washer: "washer",
} as const;