# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Car Wash Booking App API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/login:
    post:
      operationId: loginUser
      tags: [users]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginUserBody"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me:
    get:
      operationId: getCurrentUser
      tags: [users]
      summary: Get current logged in user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /carwash-owners:
    get:
      operationId: listCarwashOwners
      tags: [carwash-owners]
      summary: List all car wash businesses
      responses:
        "200":
          description: List of car wash businesses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CarwashOwner"
    post:
      operationId: createCarwashOwner
      tags: [carwash-owners]
      summary: Register a car wash business
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarwashOwnerBody"
      responses:
        "201":
          description: Car wash business created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarwashOwner"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /carwash-owners/{id}:
    get:
      operationId: getCarwashOwner
      tags: [carwash-owners]
      summary: Get car wash business by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Car wash business
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarwashOwner"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [customer, washer, owner]
        address:
          type: string
        city:
          type: string
        phone:
          type: ["string", "null"]
      required:
        - name
        - email
        - password
        - role
        - address
        - city
    LoginUserBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, washer, owner]
        address:
          type: string
        city:
          type: string
        phone:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - role
        - address
        - city
        - createdAt
    CarwashOwner:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        businessName:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        ownerName:
          type: string
        createdAt:
          type: string
      required:
        - id
        - userId
        - businessName
        - address
        - city
        - ownerName
        - createdAt
    CreateCarwashOwnerBody:
      type: object
      properties:
        businessName:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
      required:
        - businessName
        - address
        - city
...