# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Wash Booking App API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const RegisterUserBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  role: zod.enum(["customer", "cleaner", "owner"]),
  phone: zod.string().nullish(),
});

/**
 * @summary Login a user
 */
export const LoginUserBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["customer", "cleaner", "owner"]),
  phone: zod.string().nullish(),
  avatarUrl: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Get current logged in user
 */
export const GetCurrentUserResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["customer", "cleaner", "owner"]),
  phone: zod.string().nullish(),
  avatarUrl: zod.string().nullish(),
  createdAt: zod.string(),
});

/**
 * @summary Logout current user
 */
export const LogoutUserResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all car wash businesses
 */
export const ListCarwashOwnersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  businessName: zod.string(),
  phone: zod.string().nullish(),
  description: zod.string().nullish(),
  ownerName: zod.string(),
  createdAt: zod.string(),
});
export const ListCarwashOwnersResponse = zod.array(
  ListCarwashOwnersResponseItem,
);

/**
 * @summary Register a car wash business
 */
export const CreateCarwashOwnerBody = zod.object({
  businessName: zod.string(),
  phone: zod.string().nullish(),
  description: zod.string().nullish(),
});

/**
 * @summary Get car wash business by ID
 */
export const GetCarwashOwnerParams = zod.object({
  id: zod.coerce.number(),
});

export const GetCarwashOwnerResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  businessName: zod.string(),
  phone: zod.string().nullish(),
  description: zod.string().nullish(),
  ownerName: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary List all car washers
 */
export const ListWashersQueryParams = zod.object({
  available: zod.coerce.boolean().optional(),
});

export const ListWashersResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  bio: zod.string().nullish(),
  pricePerWash: zod.number(),
  available: zod.boolean(),
  rating: zod.number().nullish(),
  totalWashes: zod.number(),
  createdAt: zod.string(),
});
export const ListWashersResponse = zod.array(ListWashersResponseItem);

/**
 * @summary Register as a car washer
 */
export const CreateWasherBody = zod.object({
  phone: zod.string().nullish(),
  bio: zod.string().nullish(),
  pricePerWash: zod.number(),
});

/**
 * @summary Get washer by ID
 */
export const GetWasherParams = zod.object({
  id: zod.coerce.number(),
});

export const GetWasherResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  bio: zod.string().nullish(),
  pricePerWash: zod.number(),
  available: zod.boolean(),
  rating: zod.number().nullish(),
  totalWashes: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Get current user's washer profile
 */
export const GetMyWasherProfileResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  bio: zod.string().nullish(),
  pricePerWash: zod.number(),
  available: zod.boolean(),
  rating: zod.number().nullish(),
  totalWashes: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Update current user's washer profile
 */
export const UpdateMyWasherProfileBody = zod.object({
  phone: zod.string().nullish(),
  bio: zod.string().nullish(),
  pricePerWash: zod.number().nullish(),
  available: zod.boolean().nullish(),
});

export const UpdateMyWasherProfileResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  bio: zod.string().nullish(),
  pricePerWash: zod.number(),
  available: zod.boolean(),
  rating: zod.number().nullish(),
  totalWashes: zod.number(),
  createdAt: zod.string(),
});

const BookingStatusEnum = zod.enum([
  "searching",
  "accepted",
  "arrived",
  "in_progress",
  "declined",
  "completed",
  "cancelled",
]);

/**
 * @summary List bookings for current user
 */
export const ListBookingsQueryParams = zod.object({
  role: zod.enum(["customer", "cleaner"]).optional(),
});

export const ListBookingsResponseItem = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  washerId: zod.number().nullable(),
  customerName: zod.string(),
  washerName: zod.string().nullable(),
  customerAddress: zod.string(),
  customerLat: zod.number().nullable(),
  customerLng: zod.number().nullable(),
  scheduledAt: zod.string(),
  status: BookingStatusEnum,
  notes: zod.string().nullish(),
  vehicleType: zod.string().nullish(),
  washType: zod.string().nullish(),
  priceQuoted: zod.number(),
  createdAt: zod.string(),
  dispatchedCount: zod.number().optional(),
});
export const ListBookingsResponse = zod.array(ListBookingsResponseItem);

/**
 * @summary Create a new booking (Uber-style dispatch)
 */
export const CreateBookingBody = zod.object({
  customerAddress: zod.string(),
  customerLat: zod.number().nullish(),
  customerLng: zod.number().nullish(),
  notes: zod.string().nullish(),
  vehicleType: zod.string().nullish(),
  washType: zod.enum(["exterior", "both"]).default("exterior"),
});

/**
 * @summary Get booking by ID
 */
export const GetBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const GetBookingResponse = zod.object({
  id: zod.number(),
  customerId: zod.number(),
  washerId: zod.number().nullable(),
  customerName: zod.string(),
  customerPhone: zod.string().nullable(),
  washerName: zod.string().nullable(),
  washerPhone: zod.string().nullable(),
  customerAddress: zod.string(),
  customerLat: zod.number().nullable(),
  customerLng: zod.number().nullable(),
  washerLat: zod.number().nullable(),
  washerLng: zod.number().nullable(),
  scheduledAt: zod.string(),
  status: BookingStatusEnum,
  notes: zod.string().nullish(),
  vehicleType: zod.string().nullish(),
  washType: zod.string().nullish(),
  priceQuoted: zod.number(),
  serviceOtp: zod.string().nullable(),
  otpShared: zod.boolean(),
  createdAt: zod.string(),
  dispatchedCount: zod.number().optional(),
});

/**
 * @summary Washer accepts a booking
 */
export const AcceptBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const AcceptBookingResponse = GetBookingResponse;

/**
 * @summary Washer declines a booking
 */
export const DeclineBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const DeclineBookingResponse = GetBookingResponse;

/**
 * @summary Mark booking as complete
 */
export const CompleteBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const CompleteBookingResponse = GetBookingResponse;

/**
 * @summary Customer cancels a booking
 */
export const CancelBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const CancelBookingResponse = GetBookingResponse;