# lib/api-zod/src/generated/types/createCarwashOwnerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Car Wash Booking App API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCarwashOwnerBody {
  businessName: string;
  address: string;
  city: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  description?: string | null;
}